Tests for 'autoRefreshBot'
