+++ webpack/redux/subscribers.ts
@@ -3,7 +3,7 @@ import { Store } from "./interfaces";
 import { EnvName } from "./interfaces";
 import { all } from "../resources/selectors";
 import { SpecialStatus } from "../resources/tagged_resources";
-import { getDevice } from "../device";
+import { autoRefreshBot } from "./auto_refresh_bot";
 
 function stopThem() { return "You have unsaved work."; }
 function dontStopThem() { }
@@ -18,32 +18,16 @@ export function dontExitIfBrowserIsOnHold(state: Everything) {
   window.onbeforeunload = (unsavedWork) ? stopThem : dontStopThem;
 }
 
-interface Subscription {
+export interface Subscription {
   fn: (state: Everything) => void;
   env: EnvName;
 }
 
-/** Need to compare last state to current state. Stash old state here: */
-const mqttHistory = { last: { state: "down", at: (new Date()).toISOString() } };
-
 /** To make it easier to manage all things watching the state tree,
  * we keep subscriber functions in this array. */
 export let subscriptions: Subscription[] = [
-  {
-    env: "*",
-    fn: (s) => {
-      const x = s.bot.connectivity["bot.mqtt"];
-      const wasDown = mqttHistory.last.state === "down";
-      if (x && wasDown && x.state === "up") {
-        mqttHistory.last = x;
-        getDevice().readStatus().then(() => console.log("REFETCH"), () => { });
-      }
-    }
-  },
-  {
-    env: "*",
-    fn: dontExitIfBrowserIsOnHold
-  }
+  autoRefreshBot,
+  { env: "*", fn: dontExitIfBrowserIsOnHold }
 ];
 
 export function registerSubscribers(store: Store) {
